Unify live-play host/join and add connection resilience
One intent, one surface. Hosts express "play together" once and choose
online vs nearby inside a single sheet; guests join through one card.
A dropped connection now auto-heals instead of silently going stale.

Resilience:
- liveConnection health store (online / reconnecting / offline), separate
  from the game lifecycle so a drop never tears down the session
- relay auto-reconnect with capped backoff: rebuild the transport for the
  same code and re-announce (leader re-publishes, guest re-says hello)
- window online/offline awareness; honest "game safe" degrade
- nearby (WebRTC) reflects offline but does not fake a reconnect

Unified host:
- new PlaySheet merges the online (code / QR / link) and nearby (QR
  handshake) bodies with a mode-switch affordance
- one "Play together" button; a livebar chip shows connection health
- removes LiveShareSheet and NearbyHostSheet

Unified join:
- Home is one "Join a game" card (code input + Scan)
- the scan page routes relay links to the code join and nearby blobs to
  the handshake, so a single scanner serves both
