+++ src/lib/live/session.ts
@@ -15,7 +15,7 @@
 import { writable, derived, get } from 'svelte/store';
 import type { Participant, LiveState, LiveIntent, LiveMessage } from './protocol';
 import { PROTOCOL_VERSION } from './protocol';
-import type { SessionTransport, TransportEnvelope } from './transport';
+import type { SessionTransport, TransportEnvelope, TransportStatus } from './transport';
 import { BroadcastChannelTransport, isBroadcastSupported } from './broadcast';
 import {
   RelayTransport,
@@ -42,6 +42,14 @@ export const liveReplica = writable<LiveState | null>(null);
 export const liveError = writable<string | null>(null);
 /** True when the active session runs over the relay (cross-device) vs. same-browser only. */
 export const liveRemote = writable<boolean>(false);
+
+/**
+ * Health of the live connection, kept *separate* from {@link liveStatus} (the game lifecycle).
+ * A dropped socket never ends the game — the host stays authoritative and local-first — it just
+ * flips this to 'reconnecting' (auto-healing) or 'offline' (still retrying, game safe).
+ */
+export type LiveConnection = 'online' | 'reconnecting' | 'offline';
+export const liveConnection = writable<LiveConnection>('online');
 
 /** Convenience: a session is live when hosting or joined as a guest. */
 export const liveActive = derived(liveStatus, ($s) => $s === 'hosting' || $s === 'guest');
@@ -78,6 +86,20 @@ let peers: Participant[] = [];
 let host: HostHandlers | null = null;
 let unsubMessage: (() => void) | null = null;
 let joinTimer: ReturnType<typeof setTimeout> | undefined;
+
+// ── Reconnect state (relay only) ─────────────────────────────────────────────
+// A relay socket that drops mid-game auto-heals without tearing down the session: we rebuild
+// the transport for the same code and re-announce. Nearby (WebRTC) can't be re-signaled
+// automatically, so it reflects health but does not auto-reconnect.
+let unsubStatus: (() => void) | null = null;
+let reconnectable = false;
+let reconnectTimer: ReturnType<typeof setTimeout> | undefined;
+let reconnectAttempts = 0;
+let reconnecting = false;
+/** Set while teardown runs so a transport's own close() can't kick off a reconnect. */
+let tearing = false;
+const RECONNECT_BACKOFF_MS = [500, 1000, 2000, 4000, 6000, 8000];
+const OFFLINE_AFTER_ATTEMPTS = 4;
 
 /**
  * Serializes leader-side durable mutations. `appendRound`/`removeRound` are non-atomic
@@ -132,9 +154,11 @@ function makeTransport(code: string): SessionTransport {
   // browser profile). The engine above this line is identical either way.
   if (isRelayConfigured()) {
     liveRemote.set(true);
+    reconnectable = true;
     return new RelayTransport(effectiveRelayUrl(), code);
   }
   liveRemote.set(false);
+  reconnectable = false;
   return new BroadcastChannelTransport(code);
 }
 
@@ -164,10 +188,13 @@ export async function startHosting(code: string, handlers: HostHandlers): Promis
 
   transport = makeTransport(code);
   unsubMessage = transport.onMessage(onLeaderMessage);
+  unsubStatus = transport.onStatus(onTransportStatus);
   try {
     await transport.open();
     liveStatus.set('hosting');
+    liveConnection.set('online');
     attachUnload();
+    attachNet();
   } catch (e) {
     liveError.set(e instanceof Error ? e.message : 'Could not start live play.');
     liveStatus.set('error');
@@ -192,9 +219,11 @@ export async function joinSession(code: string, self: Participant): Promise<void
 
   transport = makeTransport(code);
   unsubMessage = transport.onMessage(onGuestMessage);
+  unsubStatus = transport.onStatus(onTransportStatus);
   try {
     await transport.open();
     attachUnload();
+    attachNet();
     send({ t: 'hello', protocol: PROTOCOL_VERSION, participant: self });
     joinTimer = setTimeout(() => {
       if (get(liveStatus) === 'connecting') {
@@ -277,10 +306,13 @@ export async function startHostingNearby(
   const rtc = new WebRtcTransport(code, 'leader');
   transport = rtc;
   liveRemote.set(true);
+  reconnectable = false;
   unsubMessage = transport.onMessage(onLeaderMessage);
+  unsubStatus = transport.onStatus(onTransportStatus);
   try {
     await transport.open(); // a leader is ready at once; guests attach as they scan in
     liveStatus.set('hosting');
+    liveConnection.set('online');
     attachUnload();
   } catch (e) {
     liveError.set(e instanceof Error ? e.message : 'Could not start nearby play.');
@@ -312,7 +344,9 @@ export async function joinSessionNearby(self: Participant): Promise<NearbyGuestC
   const rtc = new WebRtcTransport('', 'guest');
   transport = rtc;
   liveRemote.set(true);
+  reconnectable = false;
   unsubMessage = transport.onMessage(onGuestMessage);
+  unsubStatus = transport.onStatus(onTransportStatus);
 
   return {
     acceptInvite: async (offer: string): Promise<string> => {
@@ -325,6 +359,8 @@ export async function joinSessionNearby(self: Participant): Promise<NearbyGuestC
         .open()
         .then(() => {
           send({ t: 'hello', protocol: PROTOCOL_VERSION, participant: self });
+          liveConnection.set('online');
+          attachNet();
           joinTimer = setTimeout(() => {
             if (get(liveStatus) === 'connecting') {
               liveError.set('Connected, but the host didn’t respond. Ask them to try again.');
@@ -440,11 +476,120 @@ function detachUnload(): void {
   unloadAttached = false;
 }
 
+/** A session is "established" once the initial handshake has settled (not still connecting). */
+function isEstablished(): boolean {
+  const st = get(liveStatus);
+  return (role === 'leader' && st === 'hosting') || (role === 'guest' && st === 'guest');
+}
+
+/**
+ * React to transport health. The *initial* connect is owned by the entry points (which act on
+ * the open() promise while liveStatus is still 'connecting'); this only manages an already
+ * established session — flipping health and, for the relay, driving auto-reconnect.
+ */
+function onTransportStatus(status: TransportStatus): void {
+  if (tearing || !isEstablished()) return;
+  if (status === 'open') {
+    reconnectAttempts = 0;
+    liveConnection.set('online');
+  } else if (status === 'closed' || status === 'error') {
+    if (reconnectable) beginReconnect();
+    else liveConnection.set('offline');
+  }
+}
+
+/** Schedule the next reconnect attempt (deduped by the single pending timer). */
+function beginReconnect(): void {
+  if (reconnectTimer !== undefined) return;
+  liveConnection.set(reconnectAttempts >= OFFLINE_AFTER_ATTEMPTS ? 'offline' : 'reconnecting');
+  const delay = RECONNECT_BACKOFF_MS[Math.min(reconnectAttempts, RECONNECT_BACKOFF_MS.length - 1)];
+  reconnectTimer = setTimeout(() => void attemptReconnect(), delay);
+}
+
+/** Rebuild the relay transport for the same code and re-announce, without ending the session. */
+async function attemptReconnect(): Promise<void> {
+  reconnectTimer = undefined;
+  if (tearing || !reconnectable || !isEstablished()) return;
+  if (reconnecting) {
+    beginReconnect();
+    return;
+  }
+  const code = get(liveCode);
+  if (!code) return;
+  reconnecting = true;
+  reconnectAttempts++;
+  // Drop the dead transport, unsubscribing first so its own 'closed' stays silent to us.
+  unsubMessage?.();
+  unsubMessage = null;
+  unsubStatus?.();
+  unsubStatus = null;
+  try {
+    transport?.close();
+  } catch {
+    /* already closing */
+  }
+  transport = new RelayTransport(effectiveRelayUrl(), code);
+  unsubMessage = transport.onMessage(role === 'leader' ? onLeaderMessage : onGuestMessage);
+  unsubStatus = transport.onStatus(onTransportStatus);
+  try {
+    await transport.open();
+    // A reconnected leader re-publishes authoritative state; a guest re-announces to get a
+    // fresh welcome. onTransportStatus('open') flips health back to 'online'.
+    if (role === 'leader') publish();
+    else if (mySelf) send({ t: 'hello', protocol: PROTOCOL_VERSION, participant: mySelf });
+  } catch {
+    beginReconnect();
+  } finally {
+    reconnecting = false;
+  }
+}
+
+function stopReconnect(): void {
+  if (reconnectTimer !== undefined) {
+    clearTimeout(reconnectTimer);
+    reconnectTimer = undefined;
+  }
+  reconnectAttempts = 0;
+  reconnecting = false;
+}
+
+let netAttached = false;
+function onNetOnline(): void {
+  // Connectivity is back — reset backoff and retry now instead of waiting out the timer.
+  if (tearing || !reconnectable || !isEstablished()) return;
+  reconnectAttempts = 0;
+  if (reconnectTimer !== undefined) {
+    clearTimeout(reconnectTimer);
+    reconnectTimer = undefined;
+  }
+  beginReconnect();
+}
+function onNetOffline(): void {
+  if (reconnectable && isEstablished()) liveConnection.set('offline');
+}
+function attachNet(): void {
+  if (netAttached || typeof window === 'undefined') return;
+  window.addEventListener('online', onNetOnline);
+  window.addEventListener('offline', onNetOffline);
+  netAttached = true;
+}
+function detachNet(): void {
+  if (!netAttached || typeof window === 'undefined') return;
+  window.removeEventListener('online', onNetOnline);
+  window.removeEventListener('offline', onNetOffline);
+  netAttached = false;
+}
+
 async function teardown(): Promise<void> {
+  tearing = true;
+  stopReconnect();
+  detachNet();
   clearJoinTimer();
   detachUnload();
   unsubMessage?.();
   unsubMessage = null;
+  unsubStatus?.();
+  unsubStatus = null;
   transport?.close();
   transport = null;
   role = null;
@@ -456,4 +601,7 @@ async function teardown(): Promise<void> {
   seq = 0;
   opChain = Promise.resolve();
   liveRemote.set(false);
+  reconnectable = false;
+  liveConnection.set('online');
+  tearing = false;
 }
