+++ src/pages/GamePlay.svelte
@@ -25,8 +25,7 @@
   import { enableWakeLock, disableWakeLock } from '../lib/wakelock';
   import Scoreboard from '../lib/components/Scoreboard.svelte';
   import Avatar from '../lib/components/Avatar.svelte';
-  import LiveShareSheet from '../lib/components/LiveShareSheet.svelte';
-  import NearbyHostSheet from '../lib/components/NearbyHostSheet.svelte';
+  import PlaySheet from '../lib/components/PlaySheet.svelte';
   import { get } from 'svelte/store';
   import {
     startHosting,
@@ -42,6 +41,7 @@
     liveParticipants,
     liveCode,
     liveRemote,
+    liveConnection,
   } from '../lib/live/session';
   import type { HostHandlers, NearbyHostControls } from '../lib/live/session';
   import type { LiveState, LiveIntent } from '../lib/live/protocol';
@@ -262,11 +262,11 @@
   const liveSupported = isLiveSupported();
   const nearbySupported = isNearbySupported();
   let sheetOpen = $state(false);
-  let nearbyOpen = $state(false);
   let nearbyControls = $state<NearbyHostControls | null>(null);
-  // Which transport backs the active session, so the live bar reopens the right sheet.
+  // Which transport backs the active session (online = relay/broadcast, nearby = WebRTC).
   let liveKind = $state<'relay' | 'nearby' | null>(null);
   const liveLink = $derived($liveCode ? absoluteUrl(`/join/${$liveCode}`) : '');
+  const liveMode = $derived(liveKind === 'nearby' ? 'nearby' : 'online');
 
   // While hosting, route the host's own durable writes through the engine's serial queue
   // so a guest's intent can't interleave with a local save on the non-atomic append/reindex.
@@ -297,16 +297,23 @@
     return null;
   }
 
+  function hostHandlers(): HostHandlers {
+    return { self: currentSelf('leader'), buildState: buildLiveState, applyIntent: applyLiveIntent };
+  }
+
+  /** One entry point: "play together". Leads online, or nearby when the device is offline. */
+  async function startTogether() {
+    if (!game) return;
+    const offline = typeof navigator !== 'undefined' && navigator.onLine === false;
+    if (offline && nearbySupported) await startNearby();
+    else await startLive();
+  }
+
   async function startLive() {
     if (!game) return;
-    const self = currentSelf('leader');
-    const handlers: HostHandlers = {
-      self,
-      buildState: buildLiveState,
-      applyIntent: applyLiveIntent,
-    };
     try {
-      await startHosting(generateJoinCode(), handlers);
+      await startHosting(generateJoinCode(), hostHandlers());
+      nearbyControls = null;
       liveKind = 'relay';
       sheetOpen = true;
     } catch {
@@ -316,36 +323,42 @@
 
   async function startNearby() {
     if (!game) return;
-    const self = currentSelf('leader');
-    const handlers: HostHandlers = {
-      self,
-      buildState: buildLiveState,
-      applyIntent: applyLiveIntent,
-    };
     try {
-      nearbyControls = await startHostingNearby(generateJoinCode(), handlers);
+      nearbyControls = await startHostingNearby(generateJoinCode(), hostHandlers());
       liveKind = 'nearby';
-      nearbyOpen = true;
+      sheetOpen = true;
     } catch {
       showToast('Couldn’t start nearby play on this device.');
     }
   }
 
+  /** Swap transports from inside the sheet without ending the session (host's explicit choice). */
+  async function switchMode(to: 'online' | 'nearby') {
+    if (
+      $liveParticipants.length > 1 &&
+      typeof window !== 'undefined' &&
+      !window.confirm('Switching will disconnect the players who already joined. Continue?')
+    ) {
+      return;
+    }
+    if (to === 'nearby') await startNearby();
+    else await startLive();
+  }
+
   function openLiveSheet() {
-    if (liveKind === 'nearby') nearbyOpen = true;
-    else sheetOpen = true;
+    sheetOpen = true;
   }
 
   async function stopLive() {
     await leaveSession();
     sheetOpen = false;
-    nearbyOpen = false;
   }
 
   $effect(() => {
-    if (!$liveActive) {
+    // Clean up only when the session truly ends — not during a transport swap, which passes
+    // briefly through 'connecting' on its way back to 'hosting'.
+    if ($liveStatus === 'off' || $liveStatus === 'error') {
       sheetOpen = false;
-      nearbyOpen = false;
       nearbyControls = null;
       liveKind = null;
     }
@@ -387,23 +400,31 @@
     {#if $liveActive}
       <button class="card row spread livebar" onclick={openLiveSheet}>
         <span class="row" style="gap: 10px">
-          <span class="livedot" aria-hidden="true"></span>
+          <span
+            class="livedot"
+            class:reconnecting={$liveConnection === 'reconnecting'}
+            class:offline={$liveConnection === 'offline'}
+            aria-hidden="true"
+          ></span>
           <span style="display: flex; flex-direction: column; text-align: left">
             <strong>Live game</strong>
             <span class="muted" style="font-size: 0.8rem">
-              {$liveParticipants.length} here · tap to {liveKind === 'nearby' ? 'add players' : 'share'}
+              {#if $liveConnection === 'reconnecting'}
+                Reconnecting…
+              {:else if $liveConnection === 'offline'}
+                Offline · your game is safe
+              {:else}
+                {$liveParticipants.length} here · tap to {liveKind === 'nearby' ? 'add players' : 'share'}
+              {/if}
             </span>
           </span>
         </span>
         <span class="pill">{liveKind === 'nearby' ? 'Nearby' : 'Share'}</span>
       </button>
     {:else}
-      <div class="stack" style="gap: 8px; margin-top: 12px">
-        <button class="btn ghost block" onclick={startLive}>👋 Play together</button>
-        {#if nearbySupported}
-          <button class="btn ghost block" onclick={startNearby}>📡 Play nearby — no internet</button>
-        {/if}
-      </div>
+      <button class="btn ghost block" style="margin-top: 12px" onclick={startTogether}>
+        👋 Play together
+      </button>
     {/if}
   {/if}
 
@@ -501,19 +522,21 @@
   {/if}
 {/if}
 
-{#if sheetOpen && $liveActive && $liveCode}
-  <LiveShareSheet
-    code={$liveCode}
-    link={liveLink}
-    count={$liveParticipants.length}
-    remote={$liveRemote}
-    onclose={() => (sheetOpen = false)}
-    onend={stopLive}
-  />
-{/if}
-
-{#if nearbyOpen && $liveActive && nearbyControls}
-  <NearbyHostSheet controls={nearbyControls} onclose={() => (nearbyOpen = false)} onend={stopLive} />
+{#if sheetOpen && $liveActive}
+  {#key liveKind}
+    <PlaySheet
+      mode={liveMode}
+      code={$liveCode ?? ''}
+      link={liveLink}
+      remote={$liveRemote}
+      controls={nearbyControls}
+      canGoOnline={liveSupported}
+      canGoNearby={nearbySupported}
+      onclose={() => (sheetOpen = false)}
+      onend={stopLive}
+      onswitch={switchMode}
+    />
+  {/key}
 {/if}
 
 <style>
@@ -534,6 +557,29 @@
     border-radius: 999px;
     background: var(--good);
     box-shadow: 0 0 0 3px color-mix(in srgb, var(--good) 22%, transparent);
+  }
+  .livedot.reconnecting {
+    background: var(--warn);
+    box-shadow: 0 0 0 3px color-mix(in srgb, var(--warn) 22%, transparent);
+    animation: livepulse 1.1s ease-in-out infinite;
+  }
+  .livedot.offline {
+    background: var(--muted);
+    box-shadow: none;
+  }
+  @keyframes livepulse {
+    0%,
+    100% {
+      opacity: 1;
+    }
+    50% {
+      opacity: 0.3;
+    }
+  }
+  @media (prefers-reduced-motion: reduce) {
+    .livedot.reconnecting {
+      animation-duration: 3s;
+    }
   }
   .banner {
     margin-top: 12px;
